[fee-monitor] Replant the previous version of static checkBlock routine
Make a new file named "StaticChecker.ts" and replant the checking routine possible only for the static Validator set.
